Use (v)snprintf from libc when available + assiociated fixups.
git-svn-id: http://svn.enlightenment.org/svn/e/trunk@8063 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ e16/e/src/dock.c
@@ -31,7 +31,7 @@ DockIt(EWin * ewin)
    char                id[32];
 
    EDBUG(3, "DockIt");
-   Esnprintf(id, sizeof(id), "%i", ewin->client.win);
+   Esnprintf(id, sizeof(id), "%i", (unsigned)ewin->client.win);
    ac = 0;
    ic = FindItem("DEFAULT_DOCK_BUTTON", 0, LIST_FINDBY_NAME, LIST_TYPE_ICLASS);
    if (mode.dockstartx >= 0)
